Remove pooling infinite loop
