+++ Tests/MeiliSearchIntegrationTests/SettingsTests.swift
@@ -32,8 +32,6 @@ class SettingsTests: XCTestCase {
       client = try! MeiliSearch("http://localhost:7700", "masterKey")
     }
 
-    pool(client)
-
     let expectation = XCTestExpectation(description: "Create index if it does not exist")
 
     self.client.deleteIndex(UID: uid) { _ in
